@@ -4,8 +4,8 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.app.AppCompatDelegate;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 
 import ar.gob.jiaac.flysafe.R;
 
@@ -23,7 +23,7 @@ public class SplashActivity extends AppCompatActivity {
         new Handler().postDelayed(new Runnable() {
             @Override
             public void run() {
-                Intent activity = null;
+                Intent activity;
                 SharedPreferences sp = getSharedPreferences("ar.gob.jiaac.flysafe", 0);
                 if (sp.getBoolean("TermsAndConditions", false)) {
                     activity = new Intent(SplashActivity.this, HomeActivity.class);
