@@ -1,18 +1,7 @@
 package ar.gob.jiaac.flysafe.activities;
 
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatDelegate;
-import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -20,9 +9,22 @@ import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
 import com.firebase.ui.auth.AuthUI;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
+import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
@@ -77,29 +79,33 @@ public class HomeActivity extends AppCompatActivity implements HomeFragment.OnFr
         getSupportFragmentManager().addOnBackStackChangedListener(new FragmentManager.OnBackStackChangedListener() {
             @Override
             public void onBackStackChanged() {
-            if (getSupportFragmentManager().getBackStackEntryCount() > 0) {
-                getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-                toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        onBackPressed();
+                ActionBar actionBar = getSupportActionBar();
+                if (actionBar != null) {
+                    if (getSupportFragmentManager().getBackStackEntryCount() > 0) {
+                        actionBar.setDisplayHomeAsUpEnabled(true);
+                        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
+                            @Override
+                            public void onClick(View v) {
+                                onBackPressed();
+                            }
+                        });
+                    } else {
+                        actionBar.setDisplayHomeAsUpEnabled(false);
+                        toggle.syncState();
+                        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
+                            @Override
+                            public void onClick(View v) {
+                                drawerLayout.openDrawer(GravityCompat.START);
+                            }
+                        });
                     }
-                });
-            } else {
-                getSupportActionBar().setDisplayHomeAsUpEnabled(false);
-                toggle.syncState();
-                toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        drawerLayout.openDrawer(GravityCompat.START);
-                    }
-                });
-            }
+                }
             }
         });
 
         if (getSupportFragmentManager().getBackStackEntryCount() > 0) {
-            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+            ActionBar actionBar = getSupportActionBar();
+            if (actionBar != null) actionBar.setDisplayHomeAsUpEnabled(true);
             toolbar.setNavigationOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
@@ -111,29 +117,33 @@ public class HomeActivity extends AppCompatActivity implements HomeFragment.OnFr
         supportInvalidateOptionsMenu();
     }
 
+    @SuppressWarnings("StatementWithEmptyBody")
     @Override
     public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.nav_home:
-                while (getSupportFragmentManager().popBackStackImmediate());
-                replaceFragment(MainMenuFragment.newInstance(), false);
-                break;
-            case R.id.nav_stats:
-                while (getSupportFragmentManager().popBackStackImmediate());
-                replaceFragment(StatsFragment.newInstance(), false);
-                break;
-            case R.id.nav_news:
-                while (getSupportFragmentManager().popBackStackImmediate());
-                replaceFragment(WebViewFragment.newInstance("https://twitter.com/JIAAC_AR"), false);
-                break;
-            case R.id.nav_notifications:
-                while (getSupportFragmentManager().popBackStackImmediate());
-                replaceFragment(NotificationsFragment.newInstance(), false);
-                break;
-            case R.id.nav_map:
-                while (getSupportFragmentManager().popBackStackImmediate());
-                replaceFragment(ReportsMapFragment.newInstance(), false);
-                break;
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        if (fragmentManager != null) {
+            switch (item.getItemId()) {
+                case R.id.nav_home:
+                    while (fragmentManager.popBackStackImmediate()) ;
+                    replaceFragment(MainMenuFragment.newInstance(), false);
+                    break;
+                case R.id.nav_stats:
+                    while (fragmentManager.popBackStackImmediate()) ;
+                    replaceFragment(StatsFragment.newInstance(), false);
+                    break;
+                case R.id.nav_news:
+                    while (fragmentManager.popBackStackImmediate()) ;
+                    replaceFragment(WebViewFragment.newInstance("https://twitter.com/JIAAC_AR"), false);
+                    break;
+                case R.id.nav_notifications:
+                    while (fragmentManager.popBackStackImmediate()) ;
+                    replaceFragment(NotificationsFragment.newInstance(), false);
+                    break;
+                case R.id.nav_map:
+                    while (fragmentManager.popBackStackImmediate()) ;
+                    replaceFragment(ReportsMapFragment.newInstance(), false);
+                    break;
+            }
         }
         drawerLayout.closeDrawers();
         return true;
@@ -144,7 +154,7 @@ public class HomeActivity extends AppCompatActivity implements HomeFragment.OnFr
         if (requestCode == RC_SIGN_IN) {
             if (resultCode == RESULT_OK) {
                 FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-                if (!user.isEmailVerified()) {
+                if (user != null && !user.isEmailVerified()) {
                     user.sendEmailVerification().addOnCompleteListener(new OnCompleteListener<Void>() {
                         @Override
                         public void onComplete(@NonNull Task<Void> task) {
@@ -160,7 +170,9 @@ public class HomeActivity extends AppCompatActivity implements HomeFragment.OnFr
                 } else {
                     if (nv != null) {
                         TextView email = nv.getHeaderView(0).findViewById(R.id.textViewUserEmail);
-                        email.setText(user.getEmail());
+                        if (user != null) {
+                            email.setText(user.getEmail());
+                        }
                     }
                     replaceFragment(MainMenuFragment.newInstance(), false);
                 }
@@ -218,7 +230,8 @@ public class HomeActivity extends AppCompatActivity implements HomeFragment.OnFr
         }
         ft.commit();
         if (backStack) {
-            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+            ActionBar actionBar = getSupportActionBar();
+            if (actionBar != null) actionBar.setDisplayHomeAsUpEnabled(true);
         }
     }
 
