@@ -2,12 +2,14 @@ package ar.gob.jiaac.flysafe.fragments;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.text.method.LinkMovementMethod;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 
 import ar.gob.jiaac.flysafe.R;
 
@@ -19,7 +21,7 @@ public class HelpFragment extends Fragment {
 
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View v = inflater.inflate(R.layout.fragment_help, container, false);
         ((TextView) v.findViewById(R.id.userManualLabel)).setMovementMethod(LinkMovementMethod.getInstance());
