@@ -1,10 +1,9 @@
 package ar.gob.jiaac.flysafe.fragments;
 
 import android.app.Activity;
+import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,6 +11,10 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Spinner;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -37,6 +40,8 @@ public class NewAircraftFragment extends Fragment {
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View v = inflater.inflate(R.layout.fragment_new_aircraft, container, false);
+        Context context = getContext();
+        if (context == null) return v;
         editTextRegistration = v.findViewById(R.id.editTextRegistration);
         spinnerType = v.findViewById(R.id.spinnerAircraftType);
         spinnerType.setAdapter(new ArrayAdapter<>(getContext(), android.R.layout.simple_spinner_dropdown_item, aircraftTypes));
@@ -52,8 +57,12 @@ public class NewAircraftFragment extends Fragment {
                 Aircraft a = new Aircraft();
                 a.setRegistration(editTextRegistration.getText().toString().trim());
                 a.setType((AircraftType) spinnerType.getSelectedItem());
-                getTargetFragment().onActivityResult(getTargetRequestCode(), Activity.RESULT_OK, new Intent().putExtra("ar.gob.jiaac.flysafe.fragments.NewAircraftFragment", a));
-                getFragmentManager().popBackStack();
+                Fragment fragment = getTargetFragment();
+                if (fragment != null)
+                    fragment.onActivityResult(getTargetRequestCode(), Activity.RESULT_OK, new Intent().putExtra("ar.gob.jiaac.flysafe.fragments.NewAircraftFragment", a));
+                FragmentManager fragmentManager = getFragmentManager();
+                if (fragmentManager != null)
+                    fragmentManager.popBackStack();
             }
         });
         return v;
